pyproject.toml
[project]
name = "office-data-expanded"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "langchain>=1.1.0",
    "langchain-experimental>=0.4.0",
    "langchain-openai>=1.1.0",
    "openpyxl>=3.1.5",
    "pandas>=2.3.3",
    "plotly>=6.5.0",
    "python-dotenv>=1.2.1",
    "streamlit>=1.51.0",
    "tabulate>=0.9.0",
]
